feat(deprecated): introduce scaffold CRUD functionality with controllers and services for projects and workspaces, enhancing code organization and maintainability while marking existing implementations as deprecated for future removal.
