+++ server/src/_deprecated/projects.controller.ts
@@ -1,11 +1,11 @@
+import {createCrudController} from '@/_deprecated/create-crud.controller';
+import {ProjectsService} from '@/_deprecated/projects.service';
+import {ScaffoldCrudService} from '@/_deprecated/scaffold-crud.service';
 import {
     ProjectCreateDto,
     ProjectDto,
     ProjectUpdateDto
 } from '@/entities/project.entity';
-import {ProjectsService} from '@/projects/_deprecated/projects.service';
-import {createCrudController} from '@/scaffold/_deprecated/create-crud.controller';
-import {ScaffoldCrudService} from '@/scaffold/_deprecated/scaffold-crud.service';
 import {Controller} from '@nestjs/common';
 import {ApiTags} from '@nestjs/swagger';
