+++ server/src/_deprecated/projects.service.ts
@@ -1,11 +1,11 @@
+import {ScaffoldCrudDtoService} from '@/_deprecated/scaffold-crud-dto.service';
+import {ScaffoldEntityService} from '@/_deprecated/scaffold-entity.service';
 import {
     ProjectCreateDto,
     ProjectDto,
     ProjectEntity,
     ProjectUpdateDto
 } from '@/entities/project.entity';
-import {ScaffoldCrudDtoService} from '@/scaffold/_deprecated/scaffold-crud-dto.service';
-import {ScaffoldEntityService} from '@/scaffold/_deprecated/scaffold-entity.service';
 import {Injectable, Logger} from '@nestjs/common';
 import {InjectRepository} from '@nestjs/typeorm';
 import {Repository} from 'typeorm';
