+++ server/src/_deprecated/scaffold-entity.service.ts
