+++ server/src/_deprecated/scaffold-read-dto.service.ts
@@ -1,4 +1,4 @@
-import {ScaffoldEntity} from '@/scaffold/_deprecated/scaffold-entity.service';
+import {ScaffoldEntity} from '@/_deprecated/scaffold-entity.service';
 
 /**
  * @deprecated
